refactor(engine): move continuous engine tests
